Remove invalid handshake test assertion
The call to hasPending is not dependent on the isConnected call, and
the test is to ensure that handshake (e.g. connection) works, not that
messages are flushed. If we want to make some assertions on message
delivery, they should be in a different test and will need different
logic.

Fixes CASSANDRA-17979